user is prompted before each installation and setup
User can use -y as an argument to say yes to everything preemptively.

+++ ArchNolan.sh
@@ -10,7 +10,7 @@ echo "Please input your username:"
 read MAINUSER
 echo "$MAINUSER"
 
-echo "$1"
+# if the first argument is "-y" then everything is installed and configured.
 if [[ "$1" == "-y" ]]; then
         echo "Auto accepting every setup"
         AUTO=true
@@ -18,8 +18,6 @@ else
         echo "Not automatically accepting every setup. Use -y as the first argument to automatically say yes to everything\n"
         AUTO=false
 fi
-
-echo "$AUTO"
 
 # checks if the directory exists
 if [ -d "/home/$MAINUSER/" ]
@@ -32,7 +30,7 @@ else
     exit 404
 fi
 
-CWD=$(pwd)
+CWD=$(pwd) # sets the current working directory to a variable
 
 # TODO
 # yay packages
@@ -93,7 +91,6 @@ PKGS_VSCODE=(
 'ritwickdey.liveserver'
 'dsznajder.es7-react-js-snippets'
 )
-
 echo -ne "
 
 ----------------------------------------------------------------------
@@ -111,14 +108,10 @@ pacman_packages() {
         sudo pacman -S "$PKG" --noconfirm --needed
     done
 }
-
-
 if [ $AUTO == true ]
 then
-    echo "AUTO == true"
     pacman_packages
 else
-    echo "AUTO == false"
     read -p "Do you want to install Pacman Packages? " -n 1 -r
     echo    # (optional) move to a new line
     if [[ $REPLY =~ ^[Yy]$ ]]
@@ -126,11 +119,6 @@ else
         pacman_packages
     fi
 fi
-
-
-exit 404
-
-
 echo -ne "
 
 ----------------------------------------------------------------------
@@ -139,14 +127,29 @@ echo -ne "
 
 ----------------------------------------------------------------------
 
-" # installs snap packages
-sudo ln -s /var/lib/snapd/snap /snap # makes a symbolic link
+"# installs all snap packages
 
-for PKG in "${PKGS_SNAP[@]}"; do
-    echo "INSTALLING: ${PKG}"
-    sudo snap install "$PKG"
+# function to install snap packages
+snap_packages() {
+    sudo ln -s /var/lib/snapd/snap /snap # makes a symbolic link
+
+    for PKG in "${PKGS_SNAP[@]}"; do
+        echo "INSTALLING: ${PKG}"
+        sudo snap install "$PKG"
 done
 
+}
+if [ $AUTO == true ]
+then
+    snap_packages
+else
+    read -p "Do you want to install Snap Packages? " -n 1 -r
+    echo    # (optional) move to a new line
+    if [[ $REPLY =~ ^[Yy]$ ]]
+    then
+        snap_packages
+    fi
+fi
 echo -ne "
 
 ----------------------------------------------------------------------
@@ -156,9 +159,17 @@ echo -ne "
 ----------------------------------------------------------------------
 
 " # installs python pip modules
-pip install -r python_modules.txt
-
-# installs VsCode extensions
+if [ $AUTO == true ]
+then
+    pip install -r python_modules.txt
+else
+    read -p "Do you want to install Python Modules? " -n 1 -r
+    echo    # (optional) move to a new line
+    if [[ $REPLY =~ ^[Yy]$ ]]
+    then
+        pip install -r python_modules.txt
+    fi
+fi
 echo -ne "
 
 ----------------------------------------------------------------------
@@ -167,13 +178,26 @@ echo -ne "
 
 ----------------------------------------------------------------------
 
-" # installs vs code extensions
-for PKG in "${PKGS_VSCODE[@]}"; do
-    echo "INSTALLING: ${PKG}"
-    code --install-extension ${PKG}
-done
+" # installs VsCode extensions
 
-# Graphics Drivers find and install
+vs_code_extensions(){
+    for PKG in "${PKGS_VSCODE[@]}"; do
+        echo "INSTALLING: ${PKG}"
+        code --install-extension ${PKG}
+    done
+}
+
+if [ $AUTO == true ]
+then
+    vs_code_extensions
+else
+    read -p "Do you want to install VSCode Packages? " -n 1 -r
+    echo    # (optional) move to a new line
+    if [[ $REPLY =~ ^[Yy]$ ]]
+    then
+        vs_code_extensions
+    fi
+fi
 echo -ne "
 
 ----------------------------------------------------------------------
@@ -182,19 +206,33 @@ echo -ne "
 
 ----------------------------------------------------------------------
 
-" # code taken from ArchTitus, used for installing graphics drivers
-gpu_type=$(lspci)
-if grep -E "NVIDIA|GeForce" <<< ${gpu_type}; then
-    pacman -S nvidia --noconfirm --needed
-	nvidia-xconfig
-elif lspci | grep 'VGA' | grep -E "Radeon|AMD"; then
-    pacman -S xf86-video-amdgpu --noconfirm --needed
-elif grep -E "Integrated Graphics Controller" <<< ${gpu_type}; then
-    pacman -S libva-intel-driver libvdpau-va-gl lib32-vulkan-intel vulkan-intel libva-intel-driver libva-utils lib32-mesa --needed --noconfirm
-elif grep -E "Intel Corporation UHD" <<< ${gpu_type}; then
-    pacman -S libva-intel-driver libvdpau-va-gl lib32-vulkan-intel vulkan-intel libva-intel-driver libva-utils lib32-mesa --needed --noconfirm
-fi
+" # Graphics Drivers find and install
+ # code taken from ArchTitus, used for installing graphics drivers
+gpu_drivers() {
+    gpu_type=$(lspci)
+    if grep -E "NVIDIA|GeForce" <<< ${gpu_type}; then
+        pacman -S nvidia --noconfirm --needed
+        nvidia-xconfig
+    elif lspci | grep 'VGA' | grep -E "Radeon|AMD"; then
+        pacman -S xf86-video-amdgpu --noconfirm --needed
+    elif grep -E "Integrated Graphics Controller" <<< ${gpu_type}; then
+        pacman -S libva-intel-driver libvdpau-va-gl lib32-vulkan-intel vulkan-intel libva-intel-driver libva-utils lib32-mesa --needed --noconfirm
+    elif grep -E "Intel Corporation UHD" <<< ${gpu_type}; then
+        pacman -S libva-intel-driver libvdpau-va-gl lib32-vulkan-intel vulkan-intel libva-intel-driver libva-utils lib32-mesa --needed --noconfirm
+    fi
+}
 
+if [ $AUTO == true ]
+then
+    gpu_drivers
+else
+    read -p "Do you want to install GPU Drivers? " -n 1 -r
+    echo    # (optional) move to a new line
+    if [[ $REPLY =~ ^[Yy]$ ]]
+    then
+        gpu_drivers
+    fi
+fi
 echo -ne "
 
 ----------------------------------------------------------------------
@@ -204,12 +242,25 @@ echo -ne "
 ----------------------------------------------------------------------
 
 " # applies kitty config file and adds opaque theme to Kitty
-KITTY="$HOMEDIR/.config/kitty"
+kitty_config() {
+    KITTY="$HOMEDIR/.config/kitty"
 
-cp "kitty/kitty.conf" "$KITTY"
-sudo mkdir "$KITTY/themes/" # makes themes directory if it does not already exist
-cp "kitty/opaque.conf" "$KITTY/themes"
+    cp "kitty/kitty.conf" "$KITTY"
+    sudo mkdir "$KITTY/themes/" # makes themes directory if it does not already exist
+    cp "kitty/opaque.conf" "$KITTY/themes"
+}
 
+if [ $AUTO == true ]
+then
+    kitty_config
+else
+    read -p "Do you want to apply Kitty Terminal Config? " -n 1 -r
+    echo    # (optional) move to a new line
+    if [[ $REPLY =~ ^[Yy]$ ]]
+    then
+        kitty_config
+    fi
+fi
 echo -ne "
 
 ----------------------------------------------------------------------
@@ -219,10 +270,23 @@ echo -ne "
 ----------------------------------------------------------------------
 
 " # Downloads Bulk Rename Utility and installs it using wine
-# wget https://www.bulkrenameutility.co.uk/Downloads/BRU_setup.exe
-# wine BRU_setup.exe # installs Bulk Rename Utility using WINE
-# sudo rm BRU_setup.exe
+bulk_rename() {
+    wget https://www.bulkrenameutility.co.uk/Downloads/BRU_setup.exe # grabs newest version of bulkrename
+    wine BRU_setup.exe # installs Bulk Rename Utility using WINE
+    sudo rm BRU_setup.exe
+}
 
+if [ $AUTO == true ]
+then
+    bulk_rename
+else
+    read -p "Do you want to Download and Install Bulk Rename Utility? " -n 1 -r
+    echo    # (optional) move to a new line
+    if [[ $REPLY =~ ^[Yy]$ ]]
+    then
+        bulk_rename
+    fi
+fi
 echo -ne "
 
 ----------------------------------------------------------------------
@@ -232,26 +296,38 @@ echo -ne "
 ----------------------------------------------------------------------
 
 " # installs the newest version of Proton-GE and installs it to the compatabilitytools.d directory in steam
-# commented out for testing
-# curl -IkLs -o NUL -w %{url_effective} https://github.com/GloriousEggroll/proton-ge-custom/releases/latest \
-#      | grep -o "[^/]*$"\
-#      | xargs -I T \
-#        curl -kL https://github.com/GloriousEggroll/proton-ge-custom/releases/download/T/Proton-T.tar.gz \
-#        -o temp.tar.gz
-# tar xf temp.tar.gz
-# sudo rm temp.tar.gz
-# sudo rm NUL
+proton_setup() {
+    curl -IkLs -o NUL -w %{url_effective} https://github.com/GloriousEggroll/proton-ge-custom/releases/latest \
+        | grep -o "[^/]*$"\
+        | xargs -I T \
+        curl -kL https://github.com/GloriousEggroll/proton-ge-custom/releases/download/T/Proton-T.tar.gz \
+        -o temp.tar.gz
+    tar xf temp.tar.gz
+    sudo rm temp.tar.gz
+    sudo rm NUL
 
-# sudo mkdir "/home/nolan/.steam/root/compatibilitytools.d"
+    sudo mkdir "/home/nolan/.steam/root/compatibilitytools.d"
 
-# for file in *"Proton"* ; do
-#     echo "Moving $file to $HOMEDIR.steam/root/compatabilitytools.d"
-#     echo "If file is already there, deletes the file here"
-#     mv "$file" "$HOMEDIR.steam/root/compatibilitytools.d";
-#     echo "Deleting file $file"
-#     sudo rm -r "$file"; # used for if the folder is already in compatabilitytools.d
-# done
+    for file in *"Proton"* ; do
+        echo "Moving $file to $HOMEDIR.steam/root/compatabilitytools.d"
+        echo "If file is already there, deletes the file here"
+        mv "$file" "$HOMEDIR.steam/root/compatibilitytools.d";
+        echo "Deleting file $file"
+        sudo rm -r "$file"; # used for if the folder is already in compatabilitytools.d
+    done
+}
 
+if [ $AUTO == true ]
+then
+    proton_setup
+else
+    read -p "Do you want to Download and Install Proton-GE " -n 1 -r
+    echo    # (optional) move to a new line
+    if [[ $REPLY =~ ^[Yy]$ ]]
+    then
+        proton_setup
+    fi
+fi
 echo -ne "
 
 ----------------------------------------------------------------------
@@ -260,38 +336,52 @@ echo -ne "
 
 ----------------------------------------------------------------------
 
-"
+" # downloads and installs a Mac Virtual Machine. Some manual configuration is still required
+
+mac_vm() {
 # checks if the directory exists
 if [ -d "$HOMEDIR/vm/macOS-Simple-KVM" ] # TODO make this check if folder with *mac* exists
-then
-    echo "Mac VM already setup, to reinstall delete the macOS-Simple-KVM directory"
-else
-    git clone 'https://github.com/foxlet/macOS-Simple-KVM.git' 
-    mkdir "$HOMEDIR/vm"
-    for file in *"macOS"* ; do
-        cd $file # the way jumpstart.sh runs requires you to be inside the directory
-        echo "Running jumpstart.sh"
-        "./jumpstart.sh"
-        printf "\n\n\n" # prints three new lines to make the output more readable
-        echo "Adding text to basic.sh"
-        basic_line_1='-drive id=SystemDisk,if=none,file=MyDisk.qcow2 \'
-        basic_line_2='-device ide-hd,bus=sata.4,drive=SystemDisk \'
-        printf "    $basic_line_1" >> "basic.sh"
-        printf "\n    $basic_line_2" >> "basic.sh"
-        sudo qemu-img create -f qcow2 MyDisk.qcow2 30G # creates the partition for the Mac VM. Set to 32 Gigabytes
-        cd $CWD
-        echo "Moving $file to $HOMEDIR/vm/"
-        mv $file "$HOMEDIR/vm"
-        rm -r "$file" # for if the file is not moved
-        echo "Changing owners of files"
-        sudo chown "$MAINUSER" "$HOMEDIR/vm" # changes the owner of the directory to the user running the script
-        sudo chown "$MAINUSER" "$HOMEDIR/vm/$file"
-        printf "\n\nMac VM Setup with a disk size of 32 Gigabytes. Delete MyDisk.qcow2 inside $file to delete the partition."
-        printf "To create a new Disk partition for the Mac VM, run the command\n\n"
-        printf "sudo qemu-img create -f qcow2 MyDisk.qcow2 32G \n\n"
-    done
-fi
+    then
+        echo "Mac VM already setup, to reinstall delete the macOS-Simple-KVM directory"
+    else
+        git clone 'https://github.com/foxlet/macOS-Simple-KVM.git' 
+        mkdir "$HOMEDIR/vm"
+        for file in *"macOS"* ; do
+            cd $file # the way jumpstart.sh runs requires you to be inside the directory
+            echo "Running jumpstart.sh"
+            "./jumpstart.sh"
+            printf "\n\n\n" # prints three new lines to make the output more readable
+            echo "Adding text to basic.sh"
+            basic_line_1='-drive id=SystemDisk,if=none,file=MyDisk.qcow2 \'
+            basic_line_2='-device ide-hd,bus=sata.4,drive=SystemDisk \'
+            printf "    $basic_line_1" >> "basic.sh"
+            printf "\n    $basic_line_2" >> "basic.sh"
+            sudo qemu-img create -f qcow2 MyDisk.qcow2 30G # creates the partition for the Mac VM. Set to 32 Gigabytes
+            cd $CWD
+            echo "Moving $file to $HOMEDIR/vm/"
+            mv $file "$HOMEDIR/vm"
+            rm -r "$file" # for if the file is not moved
+            echo "Changing owners of files"
+            sudo chown "$MAINUSER" "$HOMEDIR/vm" # changes the owner of the directory to the user running the script
+            sudo chown "$MAINUSER" "$HOMEDIR/vm/$file"
+            printf "\n\nMac VM Setup with a disk size of 32 Gigabytes. Delete MyDisk.qcow2 inside $file to delete the partition."
+            printf "To create a new Disk partition for the Mac VM, run the command\n\n"
+            printf "sudo qemu-img create -f qcow2 MyDisk.qcow2 32G \n\n"
+        done
+    fi
+}
 
+if [ $AUTO == true ]
+then
+    mac_vm
+else
+    read -p "Do you want to Download and a Mac Virtual Machine using qemu " -n 1 -r
+    echo    # (optional) move to a new line
+    if [[ $REPLY =~ ^[Yy]$ ]]
+    then
+        mac_vm
+    fi
+fi
 
 
 #TODO setup alias for MAC to open the vm
